refactor: use vendored base58.Encode
